+++ pkg/controller/BUILD
@@ -133,7 +133,7 @@ filegroup(
         "//pkg/controller/replication:all-srcs",
         "//pkg/controller/resourcequota:all-srcs",
         "//pkg/controller/route:all-srcs",
-        "//pkg/controller/service:all-srcs",
+        "//pkg/controller/service/config:all-srcs",
         "//pkg/controller/serviceaccount:all-srcs",
         "//pkg/controller/statefulset:all-srcs",
         "//pkg/controller/testutil:all-srcs",
