+++ staging/src/k8s.io/cloud-provider/BUILD
@@ -37,6 +37,7 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//staging/src/k8s.io/cloud-provider/api:all-srcs",
+        "//staging/src/k8s.io/cloud-provider/controllers/service:all-srcs",
         "//staging/src/k8s.io/cloud-provider/fake:all-srcs",
         "//staging/src/k8s.io/cloud-provider/node:all-srcs",
         "//staging/src/k8s.io/cloud-provider/service/helpers:all-srcs",
