+++ staging/src/k8s.io/cloud-provider/controllers/service/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // Package service contains code for syncing cloud load balancers
 // with the service registry.
-package service // import "k8s.io/kubernetes/pkg/controller/service"
+package service // import "k8s.io/cloud-provider/controllers/service"
